deprecated SoccerGraphConverter, renamed to SoccerGraphConverter
